back/requirements.txt
Django>=1.11.19
django-appconf==1.0.1
djangorestframework==3.9.4
django-cors-headers==1.1.0
Pillow>=3.3.2
six==1.12.0
